move CustomDialog and ConfirmationDialog to folder base inside dialogs folder for better project structure
