+++ src/components/tables/PetitionTableWithActions.vue
@@ -81,7 +81,7 @@ import {
 import ContentApiService from '@/services/contentApiService.js';
 import PetitionTable from './PetitionTable.vue';
 import PetitionFormDialog from '@/components/dialogs/PetitionFormDialog.vue';
-import ConfirmationDialog from '@/components/dialogs/ConfirmationDialog.vue';
+import ConfirmationDialog from '@/components/dialogs/base/ConfirmationDialog.vue';
 import PetitionRevisionDialog from '@/components/dialogs/PetitionRevisionDialog.vue';
 import RoleActionButton from '../ui/RoleActionButton.vue';
